# api/openapi.yml
openapi: 3.0.3
info:
  title: GUSTAV API
  version: 0.0.2
  description: >-
    Contract-first API for minimal authentication/session integration with Keycloak.
    Endpoints cover login redirect, callback handling, logout, session info, and a
    convenience redirect for "forgot password". Sessions are represented by an
    httpOnly cookie.
servers:
  - url: /
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gustav_session
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        detail:
          type: string
...
paths:
...
  /auth/login:
    get:
      tags: [Auth]
      summary: Redirect to IdP login (Authorization Code Flow with PKCE)
      description: >-
        Redirects to the Keycloak authorization endpoint with PKCE. The server
        generates and manages the opaque `state` (CSRF) and a `nonce` (replay
        protection) for the OIDC flow; clients MUST NOT provide these values.
      operationId: startLogin
      parameters:
        - in: query
          name: redirect
          schema:
            type: string
            maxLength: 256
            pattern: '^/(?!.*//)(?!.*\\.\\.)[A-Za-z0-9._\-/]*$'
          description: >-
            Optional in-app absolute path to navigate to after login. External
            URLs are rejected; invalid values are ignored. Disallows double-slash
            and path traversal segments.
      responses:
        '302':
          description: Redirect to Keycloak authorization endpoint
          headers:
            Location:
              description: Authorization URL at Keycloak
              schema:
                type: string
  /auth/callback:
    get:
      tags: [Auth]
      summary: OAuth2/OIDC callback from Keycloak
      description: Handles the authorization code, sets session cookie, and redirects.
      operationId: authCallback
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Session established, redirect to application page
          headers:
            Set-Cookie:
              description: >-
                httpOnly session cookie `gustav_session`.
                Production: `HttpOnly; Secure; SameSite=strict`.
                Development/Test: `HttpOnly; SameSite=lax`.
                In production, `Max-Age` matches the server session TTL (default 3600 seconds).
              schema:
                type: string
            Location:
              description: In-app destination (e.g., "/" or a course page)
              schema:
                type: string
        '400':
          description: Invalid or expired code/state
          headers:
            Cache-Control:
              description: Security hardening — responses must not be cached
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    get:
      tags: [Auth]
      summary: Logout (App session + IdP end-session)
      description: >-
        Clears the application session cookie and redirects to the Identity Provider's
        end-session endpoint. After IdP logout, the browser returns to the app start page.
      operationId: logoutUnified
      parameters:
        - in: query
          name: redirect
          required: false
          schema:
            type: string
            maxLength: 256
            pattern: '^/(?!.*//)(?!.*\\.\\.)[A-Za-z0-9._\-/]*$'
            default: "/auth/logout/success"
          description: >-
            App-relative absolute path (e.g., "/courses") to redirect to after IdP logout.
            External URLs are rejected and ignored. Disallows double-slash and path traversal.
            Defaults to "/auth/logout/success". Server enforces the same regex pattern.
      responses:
        '302':
          description: Redirect to IdP end-session endpoint (then back to the app)
          headers:
            Set-Cookie:
              description: >-
                Clears the httpOnly session cookie `gustav_session` (expires it).
                Production: `HttpOnly; Secure; SameSite=strict; Max-Age=0`.
                Development/Test: `HttpOnly; SameSite=lax; Max-Age=0`.
              schema:
                type: string
            Location:
              description: IdP end-session endpoint URL
              schema:
                type: string
...
  /auth/forgot:
    get:
      tags: [Auth]
      summary: Redirect to IdP forgot password page
      description: Redirect to Keycloak password reset page.
      operationId: forgotPassword
      parameters:
        - in: query
          name: login_hint
          schema:
            type: string
            format: email
          description: Optional email to prefill on the reset page.
      responses:
        '302':
          description: Redirect to Keycloak password reset page
          headers:
            Location:
              schema:
                type: string
  /auth/register:
    get:
      tags: [Auth]
      summary: Redirect to IdP registration
      description: >-
        Redirects to Keycloak with `kc_action=register`. The server generates and
        includes a `nonce` parameter (replay protection) in the authorization request,
        analogous to `/auth/login`.
      operationId: startRegistration
      parameters:
        - in: query
          name: login_hint
          schema:
            type: string
            format: email
          description: Optional email to prefill in the registration form.
      responses:
        '302':
          description: Redirect to Keycloak registration page
          headers:
            Location:
              schema:
                type: string
...