# backend/web/routes/auth.py
"""
Authentication-related FastAPI routes (router-only module).

Why:
    Keep auth endpoints in a dedicated router to avoid duplication between the
    full app and the slim test app, and to improve maintainability.

Notes:
    - This module purposely imports from `main` inside functions to reuse the
      shared OIDC config, state/session stores, and cookie policy helpers.
      This keeps the state shared with `/auth/callback`, which remains defined
      in `main.py` for test monkeypatching compatibility.
"""

from __future__ import annotations

from fastapi import APIRouter, Request, HTTPException
from fastapi.responses import RedirectResponse, Response, HTMLResponse
from urllib.parse import urlencode
import os
import secrets

from identity_access.oidc import OIDCClient, OIDCConfig
import re
import logging


auth_router = APIRouter(tags=["Auth"])  # explicit paths, no prefix (align with OpenAPI)
logger = logging.getLogger("gustav.web.auth")

# Single source of truth for allowed in-app redirect paths
# Disallow double slashes and path traversal (".."), allow dots in names
# Keep pattern in sync with OpenAPI (api/openapi.yml)
INAPP_PATH_PATTERN = re.compile(r"^(?!.*//)(?!.*\.\.)/[A-Za-z0-9._\-/]*$")
MAX_INAPP_REDIRECT_LEN = 256


def _request_app_base(request: Request) -> str:
    """Derive the browser-facing app base from the incoming request.

    Honors trusted proxy headers when GUSTAV_TRUST_PROXY=true; otherwise uses
    ASGI's scheme/host. Returns scheme://host[:port].
    """
    import os
    trust_proxy = (os.getenv("GUSTAV_TRUST_PROXY", "false") or "").lower() == "true"
    scheme = (request.url.scheme or "http").lower()
    # Prefer ASGI-parsed host:port when not trusting proxy headers to avoid
    # Host header spoofing; fall back to Host for completeness.
    if request.url.hostname:
        if request.url.port:
            host = f"{request.url.hostname}:{request.url.port}"
        else:
            host = request.url.hostname
    else:
        host = request.headers.get("host") or ""
    if trust_proxy:
        xf_proto = (request.headers.get("x-forwarded-proto") or scheme).split(",")[0].strip()
        xf_host = (request.headers.get("x-forwarded-host") or host).split(",")[0].strip()
        scheme = (xf_proto or scheme).lower()
        host = xf_host or host
    return f"{scheme}://{host}"


def _hostport_from_url(url: str) -> str:
    """Return lowercased host[:port] from a URL string.

    Defensive parsing: falls back to empty string on errors.
    """
    try:
        from urllib.parse import urlparse

        p = urlparse(url)
        if p.hostname:
            host = p.hostname.lower()
            port = p.port
            return f"{host}:{port}" if port else host
    except Exception:
        pass
    return ""


@auth_router.get("/auth/login")
async def auth_login(request: Request, redirect: str | None = None):
    """
    Start OIDC flow with PKCE and server-side state; redirect to IdP.

    Behavior:
        - Generates code_verifier + S256 code_challenge.
        - Validates optional `redirect` to be an absolute in-app path; external
          URLs are rejected. Persists the validated redirect together with the
          server-generated state in a server-side store.
        - Redirects to Keycloak authorization endpoint.
    Permissions:
        Public.
    """
    import main  # late import to share STATE_STORE / OIDC_CFG

    code_verifier = OIDCClient.generate_code_verifier()
    code_challenge = OIDCClient.code_challenge_s256(code_verifier)
    nonce = secrets.token_urlsafe(16)
    # Security: Accept only absolute in-app paths like "/courses". Reject external URLs.
    safe_redirect = redirect if (isinstance(redirect, str) and _is_inapp_path(redirect)) else None
    rec = main.STATE_STORE.create(code_verifier=code_verifier, redirect=safe_redirect, nonce=nonce)
    final_state = rec.state
    # Use a fresh client bound to current config (allows monkeypatch in tests)
    # Prefer dynamic redirect_uri only when the current request host matches
    # the allowed app host (from WEB_BASE or configured redirect_uri).
    current_base = _request_app_base(request).rstrip("/")
    dynamic_redirect_uri = f"{current_base}/auth/callback"
    import os as _os

    allowed_base = (_os.getenv("WEB_BASE") or main.OIDC_CFG.redirect_uri).rstrip("/")
    same_host = _hostport_from_url(dynamic_redirect_uri) == _hostport_from_url(allowed_base)

    redirect_uri = dynamic_redirect_uri if same_host else main.OIDC_CFG.redirect_uri
    cfg = OIDCConfig(
        base_url=main.OIDC_CFG.base_url,
        realm=main.OIDC_CFG.realm,
        client_id=main.OIDC_CFG.client_id,
        redirect_uri=redirect_uri,
        public_base_url=main.OIDC_CFG.public_base_url,
    )
    oidc = OIDCClient(cfg)
    url = oidc.build_authorization_url(state=final_state, code_challenge=code_challenge, nonce=nonce)
    return RedirectResponse(url=url, status_code=302)


@auth_router.get("/auth/forgot")
async def auth_forgot(login_hint: str | None = None):
    """
    Redirect to Keycloak 'Forgot Password' page.
    """
    import main  # late import

    # Use browser-facing base URL if configured to avoid mixed host issues behind proxies
    public_or_internal = (main.OIDC_CFG.public_base_url or main.OIDC_CFG.base_url).rstrip("/")
    base = f"{public_or_internal}/realms/{main.OIDC_CFG.realm}/login-actions/reset-credentials"
    query = {"login_hint": login_hint} if login_hint else None
    target = f"{base}?{urlencode(query)}" if query else base
    return RedirectResponse(url=target, status_code=302)


@auth_router.get("/auth/register")
async def auth_register(request: Request, login_hint: str | None = None):
    """
    Redirect to Keycloak registration by hinting kc_action=register on the auth endpoint.

    Why:
        Keep registration on the IdP while ensuring the authorization request
        includes a fresh nonce (OIDC replay protection), same as the login flow.
    Permissions:
        Public.
    """
    import main  # late import

    code_verifier = OIDCClient.generate_code_verifier()
    code_challenge = OIDCClient.code_challenge_s256(code_verifier)
    # Phase 2: Generate nonce for replay protection and persist in state
    nonce = secrets.token_urlsafe(16)
    rec = main.STATE_STORE.create(code_verifier=code_verifier, redirect=None, nonce=nonce)
    final_state = rec.state
    # Use dynamic redirect_uri only for allowed hosts, else fallback to configured
    current_base = _request_app_base(request).rstrip("/")
    dynamic_redirect_uri = f"{current_base}/auth/callback"
    import os as _os
    allowed_base = (_os.getenv("WEB_BASE") or main.OIDC_CFG.redirect_uri).rstrip("/")
    same_host = _hostport_from_url(dynamic_redirect_uri) == _hostport_from_url(allowed_base)
    redirect_uri = dynamic_redirect_uri if same_host else main.OIDC_CFG.redirect_uri
    cfg = OIDCConfig(
        base_url=main.OIDC_CFG.base_url,
        realm=main.OIDC_CFG.realm,
        client_id=main.OIDC_CFG.client_id,
        redirect_uri=redirect_uri,
        public_base_url=main.OIDC_CFG.public_base_url,
    )
    oidc = OIDCClient(cfg)
    # Include nonce in the authorization request similar to /auth/login
    url = oidc.build_authorization_url(state=final_state, code_challenge=code_challenge, nonce=nonce)
    sep = '&' if '?' in url else '?'
    if login_hint:
        # Security: encode parameter to avoid query injection and preserve special characters
        from urllib.parse import urlencode
        url = f"{url}{sep}{urlencode({'login_hint': login_hint})}"
        sep = '&'
    url = f"{url}{sep}kc_action=register"
    return RedirectResponse(url=url, status_code=302)


@auth_router.get("/auth/logout")
async def auth_logout(request: Request, redirect: str | None = None):
    """
    Unified logout: clear app session cookie and redirect to IdP end-session.

    Behavior:
        - Deletes server-side session if present.
        - Sends Set-Cookie to expire the app session cookie.
        - Redirects (302) to Keycloak `end_session_endpoint` with
          `post_logout_redirect_uri` pointing back to the app (success page by default).
        - Accepts only in-app absolute paths for `redirect` (e.g., "/courses").
          External URLs are rejected/ignored to prevent open redirects.
    Permissions:
        Public; IdP end-session relies on IdP browser cookie.
    """
    import main  # late import for stores and cookie policy

    # Remove server-side session if present (best-effort; never fail logout)
    sid = request.cookies.get(main.SESSION_COOKIE_NAME)
    rec = None
    if sid:
        try:
            rec = main.SESSION_STORE.get(sid or "")
        except Exception as exc:
            logger.warning("Session lookup failed during logout: %s", exc.__class__.__name__)
        try:
            main.SESSION_STORE.delete(sid)
        except Exception as exc:
            logger.warning("Session delete failed during logout: %s", exc.__class__.__name__)

    # Compute IdP logout URL and app redirect target (show success banner)
    end_session = "/auth/logout/success"  # conservative fallback
    try:
        base = (main.OIDC_CFG.public_base_url or main.OIDC_CFG.base_url).rstrip("/")
        app_base = _default_app_base(main.OIDC_CFG.redirect_uri)
        # Accept only in-app absolute paths; ignore external values
        safe_redirect = redirect if (isinstance(redirect, str) and _is_inapp_path(redirect)) else None
        # After logout, go to the app success page with a re-login link
        dest = (f"{app_base}{safe_redirect}" if safe_redirect else f"{app_base}/auth/logout/success").rstrip("/")
        # Build params: prefer id_token_hint (best compatibility), else include client_id
        params = {"post_logout_redirect_uri": dest}
        if rec and getattr(rec, "id_token", None):
            params["id_token_hint"] = rec.id_token
        else:
            params["client_id"] = main.OIDC_CFG.client_id
        end_session = (
            f"{base}/realms/{main.OIDC_CFG.realm}/protocol/openid-connect/logout?" + urlencode(params)
        )
    except Exception as exc:
        logger.warning("Logout URL composition failed: %s", exc.__class__.__name__)

    resp = RedirectResponse(url=end_session, status_code=302)
    # Clear cookie consistent with environment flags
    # Late import with fallback for both package and top-level import contexts
    try:
        from ..auth_utils import cookie_opts  # type: ignore
    except Exception:  # pragma: no cover - runtime in alternative envs
        from auth_utils import cookie_opts  # type: ignore
    opts = cookie_opts(main.SETTINGS.environment)
    resp.set_cookie(
        key=main.SESSION_COOKIE_NAME,
        value="",
        httponly=True,
        secure=opts["secure"],
        samesite=opts["samesite"],
        path="/",
        expires=0,
        max_age=0,
    )
    return resp


@auth_router.get("/auth/logout/success", response_class=HTMLResponse)
async def auth_logout_success():
    """Render a minimal success page after logout with a link to /auth/login.

    Public page (allowlisted by middleware). No user data is displayed.
    """
    html = """
    <!DOCTYPE html>
    <html lang="de">
    <head>
      <meta charset="utf-8">
      <meta name="viewport" content="width=device-width, initial-scale=1" />
      <title>Abgemeldet - GUSTAV</title>
      <link rel="stylesheet" href="/static/css/gustav.css" />
    </head>
    <body class="auth-info">
      <main class="container" style="max-width: 640px; margin: 10vh auto; background: var(--color-bg-surface); padding: 24px; border: 1px solid var(--color-border); border-radius: 8px;">
        <h1>Erfolgreich abgemeldet</h1>
        <p>Du wurdest von GUSTAV und dem Anmeldedienst abgemeldet.</p>
        <p>
          <a class="button button--primary" href="/auth/login">Erneut anmelden</a>
        </p>
      </main>
    </body>
    </html>
    """
    return HTMLResponse(content=html)


# Removed separate /auth/logout/idp — unified into GET /auth/logout


# Deprecated cookie helper removed; use `auth_utils.cookie_opts` directly


def _default_app_base(redirect_uri: str) -> str:
    """Compute the absolute application base URL from a redirect URI.

    Why:
        Keycloak's end-session flow requires an absolute `post_logout_redirect_uri`.
        This helper extracts `scheme://host[:port]` from the configured
        `REDIRECT_URI`. It is tolerant to malformed values and falls back to a
        sensible local default.

    Behavior:
        - If `redirect_uri` ends with `/auth/callback`, return the prefix before it.
        - Else, return `scheme://netloc` from parsing `redirect_uri`.
        - On parsing issues, try environment fallbacks; finally use
          `http://localhost:8100`.
    """
    from urllib.parse import urlparse
    import os

    # 1) Common case: strip /auth/callback suffix
    if isinstance(redirect_uri, str) and "/auth/callback" in redirect_uri:
        prefix = redirect_uri.split("/auth/callback")[0]
        parsed = urlparse(prefix)
        if parsed.scheme and parsed.netloc:
            return f"{parsed.scheme}://{parsed.netloc}"

    # 2) Generic parse
    try:
        parsed = urlparse(redirect_uri)
        if parsed.scheme and parsed.netloc:
            return f"{parsed.scheme}://{parsed.netloc}"
    except Exception:
        pass

    # 3) Environment fallbacks
    for var in ("APP_BASE", "WEB_BASE", "REDIRECT_URI"):
        val = os.getenv(var)
        if not val:
            continue
        try:
            p = urlparse(val)
            if "/auth/callback" in val:
                base = val.split("/auth/callback")[0]
                p2 = urlparse(base)
                if p2.scheme and p2.netloc:
                    return f"{p2.scheme}://{p2.netloc}"
            if p.scheme and p.netloc:
                return f"{p.scheme}://{p.netloc}"
        except Exception:
            continue

    # 4) Last resort: safe local default for dev
    return "http://localhost:8100"


def _is_inapp_path(value: str) -> bool:
    """Return True if value is an absolute in-app path, e.g., "/", "/courses/1".

    Why:
        Prevent open redirect vulnerabilities by only allowing internal paths
        without scheme/host or query fragments. Mirrors the OpenAPI contract.
        Pattern is defined in `INAPP_PATH_PATTERN` and kept in sync with OpenAPI.
    Examples (accepted):
        "/", "/courses", "/courses/1", "/courses/list_all"
    Examples (rejected):
        "courses" (not absolute), "https://evil.com", "/a?b", "/a#b", "/.."
    """
    try:
        if not value or not isinstance(value, str):
            return False
        if len(value) > MAX_INAPP_REDIRECT_LEN:
            return False
        return bool(INAPP_PATH_PATTERN.match(value))
    except Exception:
        return False